requirements.txt
allure-behave==2.8.13
allure-python-commons==2.8.13
attrs==19.3.0
behave==1.2.6
importlib-metadata==1.6.0
parse==1.15.0
parse-type==0.5.2
pluggy==0.13.1
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
zipp==3.1.0
